requirements.txt
amqp==2.5.0
Babel==2.6.0
celery==4.3.0
certifi==2019.3.9
Django==2.1.11
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-compat==1.0.15
django-cors-headers==2.4.0
django-filter==2.1.0
django-redis==4.10.0
djangorestframework==3.9.1
mysqlclient==1.4.1
pyjwt==1.7.1
python-crontab==2.3.6
pytz==2018.9
redis==3.2.1
requests>=2.20.0
urllib3>=1.24.2
pylint==2.3.1
flower==0.9.3
python-dateutil==2.5.0
postgres==2.2.2
psycopg2-binary==2.8.2
gunicorn==19.9.0
django-extensions==2.2.1
django-nose==1.4.6
coverage==4.5.3
whitenoise==4.1.3